move prometheus logic for exporting versions to cmd package
