+++ cmd/acra-server/prometheus.go
@@ -16,6 +16,7 @@ limitations under the License.
 package main
 
 import (
+	"github.com/cossacklabs/acra/cmd"
 	"github.com/cossacklabs/acra/decryptor/base"
 	"github.com/cossacklabs/acra/utils"
 	"github.com/prometheus/client_golang/prometheus"
@@ -50,11 +51,11 @@ func registerMetrics() {
 		prometheus.MustRegister(connectionProcessingTimeHistogram)
 		base.RegisterAcraStructProcessingMetrics()
 		base.RegisterDbProcessingMetrics()
-		utils.RegisterVersionMetrics(ServiceName)
+		cmd.RegisterVersionMetrics(ServiceName)
 		version, err := utils.GetParsedVersion()
 		if err != nil {
 			panic(err)
 		}
-		utils.ExportVersionMetric(version)
+		cmd.ExportVersionMetric(version)
 	})
 }
